move labels code to own file
